+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [v1.0.0]
+
+### Changed
+
+- Use conditions for high level account and organisation level filtering on read only access
+- Can set aws_organisation_ids and/or aws_account_ids to apply the controls as fine grained as required ( down to account/org level)
+- Apply policy for service level principals as well with lambda as a default
+- Move the lifecycle policy into a data resource
+- Add outputs for details on the repository
+- Remove nested module structure
+
 ## [v0.0.5]
 
 ### Changed